Create and update core services, components, and models
This commit creates numerous files relating to the core services and
features of our Angular application. These files include the
core service ApiService and ApiService's spec for testing, several
interceptors for HTTP requests, the ApiResponse class, and various
auth-related models and services.
We also added and updated components regarding Teams, ContentLayout,
Login, and Navbar sections of the application. New views have been
included for creating and viewing teams.
A global API URL was added to environment files as well.

These changes are essential for building the app's basic
functionality and user interface.
The HTTP interceptors are used for logging HTTP requests and can be
useful for debugging and tracking.
The ApiService will be used to interact with the backend server and
is designed to be a generic service for all kinds of HTTP requests.
Auth-related services and models will manage user authentication and
user-related features in our application.